moved tests fixup
